fix: restore schema descriptions, align transaction enums, mark tool errors, tighten input validation
Supersedes #27; thanks to @jdmadison for identifying the cleared/flagColor schema mismatch.

- Re-attach property-level descriptions after the null-compatibility wrapper
  in the registry so they survive in the advertised JSON Schema instead of
  being buried inside an anyOf branch.
- Align ynab_create_transaction's cleared and flagColor with the YNAB enums
  (cleared/uncleared/reconciled and the six flag colors plus an empty string
  to clear a flag), replacing the boolean and unchecked cast. Apply the same
  empty-string flag value to ynab_update_transaction so both tools agree.
- Mark every tool failure with isError: true via a registry-level wrapper,
  so a failed write is never rendered as a successful call.
- Tighten date, limit, months, and transactionIds validation across several
  tools to match the constraints already used elsewhere in the repo.

+++ src/tools/SpendingByCategoryTool.ts
@@ -9,7 +9,7 @@ export const inputSchema = {
   budgetId: z.string().optional().describe("The ID of the budget (optional, defaults to YNAB_BUDGET_ID environment variable)"),
   sinceDate: z.string().regex(/^\d{4}-\d{2}-\d{2}$/).optional().describe("Start of the range, inclusive (ISO format: 2024-01-01). Defaults to 30 days ago."),
   untilDate: z.string().regex(/^\d{4}-\d{2}-\d{2}$/).optional().describe("End of the range, inclusive (ISO format: 2024-01-31). Defaults to no end date."),
-  limit: z.number().positive().optional().describe("Only return the top N categories by spend"),
+  limit: z.number().int().positive().optional().describe("Only return the top N categories by spend"),
 };
 
 interface SpendingByCategoryInput {
